refactor(graduate): 사이드바 export 변경
